requirements.txt
boto3
pexpect
Pillow
